+++ .github/workflows/golang-security.yml
@@ -16,4 +16,4 @@ jobs:
       - name: Run Gosec Security Scanner
         uses: securego/gosec@master
         with:
-          args: -exclude-dir=pkg/connectors/openapiclient -exclude-dir=connectors/open_policy_agent -exclude-dir=cmd -exclude-dir=test/services/pilot/mock-openapi-policymanager ./... 
+          args: -exclude-dir=pkg/connectors/openapiclient -exclude-dir=connectors/opa -exclude-dir=cmd -exclude-dir=test/services/pilot/mock-openapi-policymanager ./...
