+++ charts/fybrik/templates/opa-connector/opa-connector-hpa.yaml
@@ -4,15 +4,15 @@
 apiVersion: autoscaling/v2beta1
 kind: HorizontalPodAutoscaler
 metadata:
-  name: openpolicyagent-connector
+  name: opa-connector
   labels:
-    app.kubernetes.io/component: openpolicyagent-connector
+    app.kubernetes.io/component: opa-connector
     {{- include "fybrik.labels" . | nindent 4 }}
 spec:
   scaleTargetRef:
     apiVersion: apps/v1
     kind: Deployment
-    name: openpolicyagent-connector
+    name: opa-connector
   minReplicas: {{ .Values.opaConnector.autoscaling.minReplicas }}
   maxReplicas: {{ .Values.opaConnector.autoscaling.maxReplicas }}
   metrics:
