+++ charts/fybrik/templates/opa-connector/opa-connector-service.yaml
@@ -3,9 +3,9 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: openpolicyagent-connector
+  name: opa-connector
   labels:
-    app.kubernetes.io/component: openpolicyagent-connector
+    app.kubernetes.io/component: opa-connector
     {{- include "fybrik.labels" . | nindent 4 }}
 spec:
   type: {{ .Values.opaConnector.service.type }}
@@ -15,6 +15,6 @@ spec:
       protocol: TCP
       name: http
   selector:
-    app.kubernetes.io/component: openpolicyagent-connector
+    app.kubernetes.io/component: opa-connector
     {{- include "fybrik.selectorLabels" . | nindent 4 }}
 {{- end }}
