+++ charts/fybrik/templates/opa-connector/opa-connector-serviceaccount.yaml
@@ -6,7 +6,7 @@ kind: ServiceAccount
 metadata:
   name: {{ .Values.opaConnector.serviceAccount.name }}
   labels:
-    app.kubernetes.io/component: openpolicyagent-connector
+    app.kubernetes.io/component: opa-connector
     {{- include "fybrik.labels" . | nindent 4 }}
   {{- with .Values.opaConnector.serviceAccount.annotations }}
   annotations:
