+++ charts/fybrik/values.yaml
@@ -69,7 +69,7 @@ coordinator:
 
   # Overrides the policy manager connector URL.
   # Defaults to `<policyManager>-connector:80`.
-  policyManagerConnectorURL: "http://openpolicyagent-connector:80"
+  policyManagerConnectorURL: "http://opa-connector:80"
 
   # Configure the vault instance to be used by the coordinator manager
   vault:
@@ -411,7 +411,7 @@ opaConnector:
   connectionTimeout:
 
   # Image name or a hub/image[:tag]
-  image: "openpolicyagent-connector"
+  image: "opa-connector"
 
   # Overrides `.global.imagePullPolicy`
   imagePullPolicy: ""
@@ -425,7 +425,7 @@ opaConnector:
     # Annotations to add to the service account
     annotations: {}
     # The name of the service account to use
-    name: openpolicyagent-connector
+    name: opa-connector
 
   podAnnotations: {}
