+++ connectors/Makefile
@@ -4,7 +4,7 @@ include $(ROOT_DIR)/Makefile.env
 CONNECTORS := \
 	katalog \
 	egeria \
-	open_policy_agent
+	opa
 
 define test-target
   test:: ; cd $1 && make test
@@ -29,5 +29,5 @@ $(foreach dir,$(CONNECTORS),$(eval $(call docker-rmi-target,$(dir))))
 
 .PHONY: deploy undeploy charts
 deploy undeploy charts:
-	$(MAKE) -C open_policy_agent $@
+	$(MAKE) -C opa $@
 	$(MAKE) -C egeria $@
