+++ connectors/opa/Dockerfile
@@ -1,0 +1,7 @@
+FROM gcr.io/distroless/static:nonroot
+
+COPY opa-connector /
+
+EXPOSE 8080
+
+ENTRYPOINT ["/opa-connector"]
