+++ connectors/opa/go/model.go
