+++ connectors/opa/lib/catalog_reader.go
