+++ connectors/opa/lib/opa_reader.go
