--- connectors/open_policy_agent/Dockerfile
@@ -1,7 +1,0 @@
-FROM gcr.io/distroless/static:nonroot
-
-COPY openpolicyagent-connector /
-
-EXPOSE 8080
-
-ENTRYPOINT ["/openpolicyagent-connector"]
