+++ pipeline/pipeline.yaml
@@ -406,7 +406,7 @@ spec:
     - name: target
       value: source-build
     - name: context
-      value: fybrik/connectors/open_policy_agent
+      value: fybrik/connectors/opa
     - name: image
       value: $(params.build_image)
     - name: docker-hostname
@@ -437,15 +437,15 @@ spec:
   - name: build-image-opa
     params:
     - name: IMAGE
-      value: $(params.docker-hostname)/$(params.docker-namespace)/openpolicyagent-connector
+      value: $(params.docker-hostname)/$(params.docker-namespace)/opa-connector
     - name: DOCKERFILE
-      value: fybrik/connectors/open_policy_agent/Dockerfile
+      value: fybrik/connectors/opa/Dockerfile
     - name: TLSVERIFY
       value: "false"
     - name: BUILD_EXTRA_ARGS
       value: --build-arg REGISTRY=$(params.docker-hostname)/$(params.docker-namespace)
     - name: CONTEXT
-      value: fybrik/connectors/open_policy_agent
+      value: fybrik/connectors/opa
     - name: HELM_PROPERTY
       value: opaConnector.image
     - name: docker-hostname
@@ -475,17 +475,17 @@ spec:
       Operator: in
       Values:
       - "all"
-      - "fybrik/connectors/open_policy_agent"
-  - name: transfer-image-openpolicyagent-connector-vulnerability-advisor
+      - "fybrik/connectors/opa"
+  - name: transfer-image-opa-connector-vulnerability-advisor
     retries: 3
     taskRef:
       kind: Task
       name: skopeo-copy
     params:
     - name: srcImageURL
-      value: docker://$(params.docker-hostname)/$(params.docker-namespace)/openpolicyagent-connector@$(tasks.build-image-opa.results.IMAGE_DIGEST)
+      value: docker://$(params.docker-hostname)/$(params.docker-namespace)/opa-connector@$(tasks.build-image-opa.results.IMAGE_DIGEST)
     - name: destImageURL
-      value: docker://us.icr.io/$(params.va-scan-namespace)/openpolicyagent-connector@$(tasks.build-image-opa.results.IMAGE_DIGEST)
+      value: docker://us.icr.io/$(params.va-scan-namespace)/opa-connector@$(tasks.build-image-opa.results.IMAGE_DIGEST)
     - name: srcTLSverify
       value: "false"
     - name: destTLSverify
@@ -701,7 +701,7 @@ spec:
     - name: release_namespace
       value: $(params.NAMESPACE)
     - name: overwrite_values
-      value: blueprintNamespace=$(params.blueprintNamespace),clusterScoped=$(params.clusterScoped),global.hub=$(params.docker-hostname)/$(params.docker-namespace),global.tag="",manager.image=$(params.docker-hostname)/$(params.docker-namespace)/$(params.MANAGER_IMAGE)@$(tasks.build-image-manager.results.IMAGE_DIGEST),katalogConnector.image=$(params.docker-hostname)/$(params.docker-namespace)/$(params.KATALOG_IMAGE)@$(tasks.build-image-katalog.results.IMAGE_DIGEST),opaConnector.image=$(params.docker-hostname)/$(params.docker-namespace)/openpolicyagent-connector@$(tasks.build-image-opa.results.IMAGE_DIGEST),cluster.name=$(params.NAMESPACE),coordinator.vault.address=http://$(params.NAMESPACE)-vault.$(params.NAMESPACE):8200,vault.address=http://$(params.NAMESPACE)-vault.$(params.NAMESPACE):8200,$(params.fybrik-values)
+      value: blueprintNamespace=$(params.blueprintNamespace),clusterScoped=$(params.clusterScoped),global.hub=$(params.docker-hostname)/$(params.docker-namespace),global.tag="",manager.image=$(params.docker-hostname)/$(params.docker-namespace)/$(params.MANAGER_IMAGE)@$(tasks.build-image-manager.results.IMAGE_DIGEST),katalogConnector.image=$(params.docker-hostname)/$(params.docker-namespace)/$(params.KATALOG_IMAGE)@$(tasks.build-image-katalog.results.IMAGE_DIGEST),opaConnector.image=$(params.docker-hostname)/$(params.docker-namespace)/opa-connector@$(tasks.build-image-opa.results.IMAGE_DIGEST),cluster.name=$(params.NAMESPACE),coordinator.vault.address=http://$(params.NAMESPACE)-vault.$(params.NAMESPACE):8200,vault.address=http://$(params.NAMESPACE)-vault.$(params.NAMESPACE):8200,$(params.fybrik-values)
     - name: release_version
       value: 0.1.0
     - name: image
