+++ test/services/pilot/mock-openapi-policymanager/log.sh
@@ -6,5 +6,5 @@ set -x
 POD=$(kubectl get pod -l app=manager-client -n fybrik-system -o jsonpath="{.items[0].metadata.name}")
 kubectl logs -n fybrik-system $POD >out-managerclient.log
 
-POD=$(kubectl get pod -l app.kubernetes.io/component=openpolicyagent-connector -n fybrik-system -o jsonpath="{.items[0].metadata.name}")
+POD=$(kubectl get pod -l app.kubernetes.io/component=opa-connector -n fybrik-system -o jsonpath="{.items[0].metadata.name}")
 kubectl logs -n fybrik-system --follow $POD
